SmErRecommendation-master-v3/requirements.txt
Django==5.2
djangorestframework==3.16.0
textblob==0.19.0
numpy==1.26.4
pandas==2.2.3
python-dateutil==2.9.0.post0
pytz==2024.2
nltk==3.9.1
regex==2024.11.6
asgiref==3.8.1
sqlparse==0.5.3
PyQt5==5.15.11
PyQtWebEngine==5.15.7
SQLAlchemy==2.0.25
Werkzeug==3.0.1
surprise==0.1
pillow==11.2.1